Continue building shopping cart screen
This commit continues the shopping cart build. The shopping cart screen now includes funcionality to include the items in the cart with selectable quantities, number of items in cart, subtotal, and proceed to checkout button which will eventually redirect the user to either the shipping page if logged in, or the login screen if they are not logged in.