Add custom authentication middleware
This commit adds the custom authentication middleware. The token that is generated for a user upon succesful authentication contains user id , and we're getting the user ID and fetching the user in the  middleware and then assigning it to request.user. Now we can use request.user in any protected route that we want.