Add order screen
This commit adds the OrderScreen.js to the project. The order details are gathered via getOrderDetails which takes in the orderID and then if the order ID matches an ID in the database it is then brought via the backend controller and routes to flow through as state to the order details reducer and then we are given access to the data once the actions are imported and dispatched in the Order screen.