Add backend functionality to update payment status
This commit adds the protected route '/:id/pay' to orderRoutes.js and then adds updateOrderToPaid in orderControllers.js. updateOrderToPaid takes the current order (if there is one) and returns a new variable called updatedOrder as json data. updatedOrder sets isPaid to true, paidAt to Date.now(), and paymentResult to an object of information about the order. Throws error if no order found or error