Add delete user functionality for admin
This commit adds the functionality to delete a user. On the backend we add the controller for deleteUser, add route for deleteUser, and we verified user is deleted correctly when sending the DELETE request from postman.

On frontend we add the Constant and reducer for delete user, add it to the store.js, create action for deleteUser, and then add the deleteUser functionality in the UserListScreen.