Add create and update product endpoints
This commit adds controller functions and routes for creating a product and editing a product. When we click create product it will create a product right away with some sample data, then we want to be taken to a page that will allow us to edit the product information. Verified sample product is being created correctly via PostMan.