Add Caoursel component to home page
This commit adds the carousel to enable to show the top rated products..Created controller and route in backend to getTopProducts. Add constants for PRODUCT TOP, added Reducer for productTopRatedReducer, added productTopReducer in store.js in imports and reducer productTopRated: productTopRatedReducer and action productTop. Adds carousel to home screen and styles component.